refactor(agent): move file, image and frame source adapters into agent/infrastructure
Move-only: FileServiceImpl, ImageService and DirectoryFrameSource implement
agentdomain's FileService, ImageService and FrameSource ports. They are
infrastructure adapters for the agent context and now sit next to the tool
formatter that was already there.

This empties the last of the loose files from internal/services; the container
drops its services import for agentinfra.
