+++ internal/agent/infrastructure/file.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"fmt"
