+++ internal/agent/infrastructure/file_test.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"os"
