+++ internal/agent/infrastructure/frame_source_directory.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"bytes"
