+++ internal/agent/infrastructure/frame_source_directory_test.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"os"
