+++ internal/agent/infrastructure/image_service.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"bytes"
