+++ internal/agent/infrastructure/image_service_test.go
@@ -1,4 +1,4 @@
-package services
+package infrastructure
 
 import (
 	"bytes"
